+++ cpp/wrapper.hpp
@@ -258,13 +258,13 @@ SpvResult spv_compiler_set_member_decoration(SpvCompiler *compiler,
 
 SpvResult spv_compiler_get_declared_struct_size(const SpvCompiler *compiler,
                                                 const uint32_t id,
-                                                uint32_t *result,
+                                                size_t *result,
                                                 SpvDString *error_msg);
 
 SpvResult spv_compiler_get_declared_struct_member_size(const SpvCompiler *compiler,
                                                        const uint32_t id,
                                                        const uint32_t index,
-                                                       uint32_t *result,
+                                                       size_t *result,
                                                        SpvDString *error_msg);
 
 SpvResult spv_compiler_rename_interface_variable(SpvCompiler *compiler,
